Client doesn't want chapters, and wants footnotes to be near their source, not at bottom of page.
This is a challenge as MSWord doesn't allow much flexibility in its exporting to HTML process.

With this current code, I am transforming the text of the footnotes into a tooltip that appears when the reader hovers over the footnote number in the body text. If client doesn't approve of this, I will use a react component for it.